backend-django/requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis==5.4.0
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
bcrypt==4.2.0
